Fix: Improve user table by migrate from integer data type as user id to bigInteger to handle the bigger version of the telegram user id

+++ tg_bot/modules/sql/users_sql.py
@@ -2,7 +2,7 @@ import threading
 
 from sqlalchemy import (
     Column,
-    Integer,
+    BigInteger,
     UnicodeText,
     String,
     ForeignKey,
@@ -16,7 +16,7 @@ from tg_bot.modules.sql import BASE, SESSION
 
 class Users(BASE):
     __tablename__ = "users"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     username = Column(UnicodeText)
 
     def __init__(self, user_id, username=None):
